+++ src/wrapper/QueryRewriter.cpp
@@ -120,7 +120,7 @@ void QueryRewriter::applyAnalyzer() {
     // now erase the data of erased keywords
     std::vector<std::string> rawQueryKeywords;
     std::vector<float> keywordFuzzyLevel;
-    std::vector<float> keywordBoostLevel;
+    std::vector<unsigned> keywordBoostLevel;
     std::vector<srch2is::TermType> keywordPrefixComplete;
     std::vector<std::vector<std::string> > fieldFilter;
     std::vector<srch2is::BooleanOperation> fieldFilterOps;
